+++ packages/core-js/proposals/url.js
@@ -2,3 +2,4 @@
 require('../modules/web.url');
 require('../modules/web.url.to-json');
 require('../modules/web.url-search-params');
+require('../modules/web.url-search-params.size');