+++ packages/core-js/web/url-search-params.js
@@ -1,4 +1,5 @@
 require('../modules/web.url-search-params');
+require('../modules/web.url-search-params.size');
 var path = require('../internals/path');
 
 module.exports = path.URLSearchParams;